raspi_zzanggu/requirements.txt
SpeechRecognition
pvrecorder
numpy
pvporcupine
python-dotenv
langchain
langchain-openai
langchain_core
sounddevice
soundfile